feat: add argument forwarding for all scripts
This commit adds the hidden argument "--" implicitly exported by all
scripts.

When the parser sees the runtime argument "--" it will short-circuit the
next iteration of the runtime arguments loop, feeding all runtime args
into an array variable called "$forwarded", available to script authors.

This provides a more natural way to forward arguments from a script into
another command.

For example, if you're writing a kubectl wrapper you can have a script
which forwards arguments directly to the kubectl cli:

$> cmds k8s ctl.sh -- get pods

Inside the script you can use the $forwarded and simply pass it to the
cli command:

execute() {
    kubectl $forwarded
}

The alternative to this approach is a "multi-word" string value to an
argument. For example

$> cmds k8s ctl.sh --cmd "get pods"

The dowside to this is "get pods" is a subtly a single element all the
way do to the script.

Therefore,

execute() {
    kubectl $cmd
}

Does not work as expected, this winds up passing the "multi-word" string
value directly to kubectl behind the scenes, e.g. "kubectl 'get pods'".

Kubectl will think you're asking for a command called 'get pods' and not
asking providing a subcommand->command pair such as "get" "pods".

One can make the above work, if they desire, but explicitly performing
word splitting:

execute() {
    kubectl ${(z)cmd}
}

This expands $cmd but splits the multi-word string into individual
elements.

However, at this point using "--" becomes more intuitive.

Signed-off-by: ldelossa <louis.delos@gmail.com>